+++ core/src/mantle/ops/channel/mod.rs
@@ -5,7 +5,12 @@ pub mod withdraw;
 
 use std::fmt::{Display, Formatter};
 
-use crate::utils::serde_bytes_newtype;
+use nom::IResult;
+
+use crate::{
+    mantle::nom::{NomDecode, NomEncode},
+    utils::serde_bytes_newtype,
+};
 
 pub type ChannelKeyIndex = u16;
 
@@ -20,6 +25,19 @@ impl Display for ChannelId {
     }
 }
 
+impl NomEncode for ChannelId {
+    fn encode(&self) -> Vec<u8> {
+        self.0.encode()
+    }
+}
+
+impl NomDecode for ChannelId {
+    fn decode(bytes: &[u8]) -> IResult<&[u8], Self> {
+        let (bytes, inner) = <[u8; _]>::decode(bytes)?;
+        Ok((bytes, Self::from(inner)))
+    }
+}
+
 /// The id of the previous message in the channel
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
 pub struct MsgId([u8; 32]);
@@ -29,6 +47,19 @@ impl Display for MsgId {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         let hex_string = hex::encode(self.0);
         write!(f, "{hex_string}")
+    }
+}
+
+impl NomEncode for MsgId {
+    fn encode(&self) -> Vec<u8> {
+        self.0.encode()
+    }
+}
+
+impl NomDecode for MsgId {
+    fn decode(bytes: &[u8]) -> IResult<&[u8], Self> {
+        let (bytes, inner) = <[u8; _]>::decode(bytes)?;
+        Ok((bytes, Self::from(inner)))
     }
 }
